Factor out ConsentOrder
