+++ identity/app/utils/ConsentOrder.scala
@@ -1,0 +1,45 @@
+package utils
+
+import com.gu.identity.model.{Consent, User}
+
+/** Utility for changing the order of consents */
+object ConsentOrder {
+
+  /** This list controls the order. */
+  private val orderedConsentIds: List[String] =
+    List(
+      "supporter",
+      "jobs",
+      "holidays",
+      "events",
+      "offers"
+    )
+
+  /**
+    * Ordered + optionally hintend consents
+    *
+    * @param userDO user object from identity-model which has consents field
+    * @param consentHint optional hint which would move that particular consent to the front
+    * @return copy of user with reordered consents
+    */
+  def userWithOrderedConsents(userDO: User, consentHint: Option[String]): User =
+    userDO.copy(consents = hintedConsents(orderedConsents(userDO.consents), consentHint))
+
+  /** If consentHint is provided it moves that consent to the head of consents list */
+  private def hintedConsents(consents: List[Consent], consentHint: Option[String]): List[Consent] = {
+
+    // https://stackoverflow.com/questions/24870729/moving-an-element-to-the-front-of-a-list-in-scala
+    def moveToFront(hint: String, consents: List[Consent]): List[Consent] = {
+      consents.span(consent => consent.id != hint) match {
+        case (as, h :: bs) => h :: as ++ bs
+        case _ => consents
+      }
+    }
+
+    consentHint.map(hint => moveToFront(hint, consents)).getOrElse(consents)
+  }
+
+  private def orderedConsents(consents: List[Consent]): List[Consent] =
+    orderedConsentIds.map(id => consents.find(_.id == id)).flatten
+
+}
